fixed based on feedback
